Refactor HTTP host routes into explicit contract groups
