+++ internal/server/routes_bindings.go
@@ -1,0 +1,39 @@
+package server
+
+import (
+	"log"
+
+	"github.com/go-chi/chi/v5"
+)
+
+func (s *Server) mountBindingRoutes(r chi.Router) {
+	if s.bindings == nil {
+		return
+	}
+	for _, name := range s.bindings.List() {
+		binding, err := s.bindings.Get(name)
+		if err != nil {
+			log.Printf("warning: skipping binding %q routes: %v", name, err)
+			continue
+		}
+		for _, route := range binding.Routes() {
+			handler := route.Handler
+			if !route.Public {
+				if route.ProxyAuth {
+					handler = s.proxyAuthMiddleware(handler)
+				} else {
+					handler = s.authMiddleware(handler)
+				}
+			}
+			if route.Connect {
+				if s.connectHandler != nil {
+					log.Printf("warning: binding %q registers CONNECT but another binding already claimed it; skipping", name)
+					continue
+				}
+				s.connectHandler = handler
+				continue
+			}
+			r.Method(route.Method, "/bindings/"+name+route.Pattern, handler)
+		}
+	}
+}
