+++ internal/server/routes_user.go
@@ -1,0 +1,36 @@
+package server
+
+import "github.com/go-chi/chi/v5"
+
+func (s *Server) mountAuthenticatedRoutes(r chi.Router) {
+	r.Group(func(r chi.Router) {
+		r.Use(s.authMiddleware)
+
+		r.Get("/integrations", s.listIntegrations)
+		r.Delete("/integrations/{name}", s.disconnectIntegration)
+		r.Get("/integrations/{name}/operations", s.listOperations)
+		r.Get("/runtimes", s.listRuntimes)
+		r.Get("/bindings", s.listBindings)
+
+		r.Get("/{integration}/{operation}", s.executeOperation)
+		r.Post("/{integration}/{operation}", s.executeOperation)
+
+		r.Post("/auth/start-oauth", s.startIntegrationOAuth)
+		r.Post("/auth/connect-manual", s.connectManual)
+
+		r.Get("/connections/staged/{id}", s.getStagedConnection)
+		r.Post("/connections/staged/{id}/select", s.selectStagedConnection)
+		r.Delete("/connections/staged/{id}", s.cancelStagedConnection)
+
+		r.Post("/tokens", s.createAPIToken)
+		r.Get("/tokens", s.listAPITokens)
+		r.Delete("/tokens/{id}", s.revokeAPIToken)
+
+		r.Post("/egress-clients", s.createEgressClient)
+		r.Get("/egress-clients", s.listEgressClients)
+		r.Delete("/egress-clients/{id}", s.deleteEgressClient)
+		r.Post("/egress-clients/{id}/tokens", s.createEgressClientToken)
+		r.Get("/egress-clients/{id}/tokens", s.listEgressClientTokens)
+		r.Delete("/egress-clients/{id}/tokens/{tokenID}", s.revokeEgressClientToken)
+	})
+}
